+++ examples/window.rs
@@ -8,7 +8,6 @@ fn main() {
         .build(&events_loop)
         .unwrap();
 
-    _window.center();
 
     events_loop.run_forever(|event| {
         println!("{:?}", event);