+++ src/platform/windows/window.rs
@@ -728,11 +728,11 @@ unsafe fn init(
         let window_size = win.get_outer_size();
 
         let monitor;
-        if attributes.start_monitor == -1 {
+        if attributes.monitor_index == -1 {
             monitor = win.get_primary_monitor();
         } else {
-            if attributes.start_monitor < win.get_available_monitors().len() as i16 {
-                monitor = win.get_available_monitors()[attributes.start_monitor as usize].clone();
+            if attributes.monitor_index < win.get_available_monitors().len() as i16 {
+                monitor = win.get_available_monitors()[attributes.monitor_index as usize].clone();
             } else {
                 monitor = win.get_primary_monitor();
             }