+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
-- Added in WindowAttributes `center_window: bool` `start_monitor: i16`, Window centering is by default enabled on windows.
-- Added convenience functions to manage window centering, `center()` ,`set_center_monitor()`, `set_centered()` see more in window struct.
+- Added to `WindowAttributes` the attributes `center_window: bool` and `monitor_index: i16`. Window centering is by default enabled.
+- Added the following convenience functions for managing window centering: center(), set_target_monitor(), and set_centered(). Refer to the documentation for how to use them.
 # Version 0.19.1 (2019-04-08)
 
 - On Wayland, added a `get_wayland_display` function to `EventsLoopExt`.